fix heroku deployment
